refactor: extract mongodb parser to proximadb-mongodb-parser crate
Extracts src/query/parsers/mongodb.rs (2,008 LOC) into a dedicated
proximadb-mongodb-parser crate at crates/binding/.

- New crate depends on proximadb-proto (foundation) for proto types, not
  the root crate — avoids a build cycle and keeps it a leaf-leaning crate.
- Root keeps a re-export shim (src/query/parsers/mongodb.rs) that glob
  re-exports the crate surface, so all existing callers using
  proximadb::query::parsers::mongodb::* resolve unchanged.
- The two root-local trait impls (QueryParser for MongoDBParser, ToFilter
  for MongoDBExpression) move to the shim, since they bind extracted types
  to traits shared with the Cypher parser that remain in the root.
- 1 dep path swapped (crate::proto -> proximadb_proto). 15 unit tests
  move with the file and pass.
- Registered in workspace members, root [dependencies], and CI feature-matrix.
