+++ .github/workflows/ci.yml
@@ -755,6 +755,7 @@ jobs:
           # crate compiles against root at every tier. Python-feature + wheel build
           # is a qa→main (FULL) concern (heavy pyo3/numpy deps).
           - { label: "embedded",             pkg: proximadb-embedded, args: "" }
+          - { label: "mongodb-parser",       pkg: proximadb-mongodb-parser, args: "" }
     steps:
       - uses: actions/checkout@v4
       - uses: dtolnay/rust-toolchain@1.95.0
