+++ Cargo.toml
@@ -120,6 +120,7 @@ members = [
     "crates/query/proximadb-uql",
     "crates/query/proximadb-vector-query",
     "crates/binding/proximadb-spark-jni",
+    "crates/binding/proximadb-mongodb-parser",
 ]
 
 [workspace.package]
@@ -603,6 +604,7 @@ moka = { version = "0.12", features = ["future"] }
 proximadb-raptor-common = { version = "0.2.0", path = "crates/storage/proximadb-raptor-common" }
 proximadb-raptor-engine = { version = "0.2.0", path = "crates/storage/proximadb-raptor-engine" }
 proximadb-engine-core = { version = "0.2.0", path = "crates/storage/proximadb-engine-core" }
+proximadb-mongodb-parser = { version = "0.2.0", path = "crates/binding/proximadb-mongodb-parser" }
 
 # GPU acceleration dependencies (platform-specific)
 [target.'cfg(all(target_os = "macos", target_arch = "aarch64"))'.dependencies]
