+++ crates/binding/proximadb-mongodb-parser/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "proximadb-mongodb-parser"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "MongoDB query language parser for ProximaDB — parses MongoDB-style queries, projections, and aggregation pipelines into DocumentFilter."
+
+[lib]
+name = "proximadb_mongodb_parser"
+path = "src/lib.rs"
+
+[dependencies]
+# Depends on the foundation proto crate for proto types (one-directional, no cycle
+# with the root crate). The root `proximadb` crate re-exports these via its `proto`
+# module, but depending on `proximadb-proto` directly keeps this a leaf-leaning crate
+# and lets the root depend on us for the re-export shim without forming a build cycle.
+proximadb-proto = { path = "../../foundation/proximadb-proto" }
+anyhow.workspace = true
+nom = "7"
+serde_json.workspace = true
+
+[lints]
+workspace = true
